- 调整代码

+++ Engine/Renderer/GLES3/src/GLES3Mesh.cpp
@@ -186,7 +186,7 @@ uint32_t CGLES3Mesh::GetIndexCount(int indexDraw) const
 uint32_t CGLES3Mesh::GetIndexOffset(int indexDraw) const
 {
 	if (indexDraw >= 0 && indexDraw < (int)m_draws.size() && m_pIndexBuffer) {
-		switch (m_pIndexBuffer->GetIndexType()) {
+		switch ((int)m_pIndexBuffer->GetIndexType()) {
 		case GL_UNSIGNED_BYTE:  return m_draws[indexDraw].firstIndex;
 		case GL_UNSIGNED_SHORT: return m_draws[indexDraw].firstIndex * 2;
 		case GL_UNSIGNED_INT:   return m_draws[indexDraw].firstIndex * 4;
